Fix all test with extends AbstractTest.
