+++ pom.xml
@@ -40,7 +40,16 @@
             <artifactId>calculator</artifactId>
             <version>1.0</version>
             <scope>system</scope>
-            <systemPath>${project.basedir}/src/main/java/ru/training/at/hw1/calculator-1.0.jar</systemPath>
+            <systemPath>${project.basedir}/src/main/resources/calculator-1.0.jar</systemPath>
+            <!--
+            [WARNING] Some problems were encountered while building
+            the effective model for ru.training.at:rafael_reziapov:jar:1.0-SNAPSHOT
+            [WARNING] 'dependencies.dependency.systemPath'
+            for com.epam.tat.module4:calculator:jar
+            should not point at files within the project directory,
+            ${project.basedir}/src/main/resources/calculator-1.0.jar
+            will be unresolvable by dependent projects @ line 43, column 25
+            -->
             <type>jar</type>
             <optional>true</optional>
         </dependency>
