+++ src/test/java/ru/training/at/hw1/calculator/CalculatorAddTest.java
@@ -1,27 +1,10 @@
 package ru.training.at.hw1.calculator;
 
-import com.epam.tat.module4.Calculator;
 import org.testng.annotations.*;
-import org.testng.asserts.SoftAssert;
 import ru.training.at.hw1.dataproviders.SumDoubleDataPositiveTestProvider;
 import ru.training.at.hw1.dataproviders.SumLongDataPositiveTestProvider;
 
-public class CalculatorAddTest {
-    Calculator calculator;
-    SoftAssert softAssert;
-
-    @BeforeMethod(groups = {"add_subtract"})
-    public void setUp() {
-        calculator = new Calculator();
-        softAssert = new SoftAssert();
-    }
-
-    @AfterMethod(groups = {"add_subtract"})
-    public void tearDown() {
-        calculator = null;
-        softAssert = null;
-    }
-
+public class CalculatorAddTest extends AbstractTest {
     @Test(groups = {"add_subtract"}, priority = 10,
                 dataProvider = "SumDoubleDataPositiveTestProvider",
                 dataProviderClass = SumDoubleDataPositiveTestProvider.class)
@@ -37,5 +20,4 @@ public class CalculatorAddTest {
         long actual = calculator.sum(a, b);
         softAssert.assertEquals(actual, expected, "SumLongDataPositiveTest is wrong");
     }
-
 }
