+++ src/test/java/ru/training/at/hw1/calculator/CalculatorDivideTest.java
@@ -1,10 +1,6 @@
 package ru.training.at.hw1.calculator;
 
-import com.epam.tat.module4.Calculator;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import org.testng.asserts.SoftAssert;
 import ru.training.at.hw1.dataproviders.DivideByZeroDataTestProvider;
 import ru.training.at.hw1.dataproviders.DivideDoubleDataPositiveTestProvider;
 import ru.training.at.hw1.dataproviders.DivideLongDataPositiveTestProvider;
@@ -12,21 +8,7 @@ import ru.training.at.hw1.dataproviders.DivideLongDataPositiveTestProvider;
 import static org.testng.AssertJUnit.assertEquals;
 import static org.testng.AssertJUnit.fail;
 
-public class CalculatorDivideTest {
-    Calculator calculator;
-    SoftAssert softAssert;
-
-    @BeforeMethod(groups = {"multiply_divide"})
-    public void setUp() {
-        calculator = new Calculator();
-        softAssert = new SoftAssert();
-    }
-
-    @AfterMethod(groups = {"multiply_divide"})
-    public void tearDown() {
-        calculator = null;
-        softAssert = null;
-    }
+public class CalculatorDivideTest extends AbstractTest {
 
     @Test(groups = {"multiply_divide"}, priority = 70,
             expectedExceptionsMessageRegExp = "Attempt to divide by zero",
