+++ src/test/java/ru/training/at/hw1/calculator/CalculatorMultiplyTest.java
@@ -1,28 +1,10 @@
 package ru.training.at.hw1.calculator;
 
-import com.epam.tat.module4.Calculator;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-import org.testng.asserts.SoftAssert;
 import ru.training.at.hw1.dataproviders.MultiplyDoubleDataPositiveTestProvider;
 import ru.training.at.hw1.dataproviders.MultiplyLongDataPositiveTestProvider;
 
-public class CalculatorMultiplyTest {
-    Calculator calculator;
-    SoftAssert softAssert;
-
-    @BeforeMethod(groups = {"multiply_divide"})
-    public void setUp() {
-        calculator = new Calculator();
-        softAssert = new SoftAssert();
-    }
-
-    @AfterMethod(groups = {"multiply_divide"})
-    public void tearDown() {
-        calculator = null;
-        softAssert = null;
-    }
+public class CalculatorMultiplyTest extends AbstractTest {
 
     @Test(groups = {"multiply_divide"}, priority = 50,
             dataProvider = "MultiplyLongDataPositiveTestProvider",
